Did some Cleanup in tower targeting SO
I was trying to inherit form the targetingSo parsa made but the way it has been made, it is not reusable for the spells. So I need to make one. I took the opportunity to do some clean up in the code correcting  english mistakes and removing unused using directive.