AbilityState Refactor, PlayerGroundTargetStateSo, Spell UI, TowerBuild Ability
Change event call in abilityso to heppen in each state,
created the tower ability so,
ability targeting state now need a targeting state so,
created a manager and script for spell UI (transparent tower), also change prefab accordingly,
created a ground targeting state so, also did the Scriptable object of it.
created a bug in the factory where postInit function of objects was never called.
change IState interface to an abstract State class.
corrected various bug of null reference exception when trying to create a tower in game.