Fix turret flashing again
I tought the issue was gone but it came back, it was not releasing properly to set the is press bool to false if not using the same input... i think,? but i know it has issue with the isPress bool staying true for some reason.

deleted isPress bool in abilitySo and change the transition condition from ready to target to check if isSelected.